closes #87 Divide Budget Summary/Detail components into relevant subcomponents and stub them out
