+++ src/routes.js
@@ -26,9 +26,17 @@ import DevelopmentSummary from './topics/development/summary/DevelopmentSummary'
 import DevelopmentDetail from './topics/development/detail/DevelopmentDetail';
 import DevelopmentSLADashboard from './topics/development/sla_dashboard/SLADashboard';
 import CrimeSummary from './topics/crime/summary/CrimeSummary';
-import BudgetSummary from './topics/budget/summary/BudgetSummary';
 import CrimeDetail from './topics/crime/detail/CrimeDetail';
 import Maintenance from './topics/maintenance/Maintenance';
+// Budget
+import BudgetSummary from './topics/budget/summary/BudgetSummary';
+import BudgetSummaryDetails from './topics/budget/summary/BudgetSummaryDetails';
+// for stand-alone budget iframings
+import BudgetDetailsTable from './topics/budget/summary/BudgetDetailsTable';
+import BudgetDetailsTreemap from './topics/budget/summary/BudgetDetailsTreemap';
+import SummaryUse from './topics/budget/summary/SummaryUse';
+import SummaryDepartments from './topics/budget/summary/SummaryDepartments';
+import SummaryCashFlow from './topics/budget/summary/SummaryCashFlow';
 
 const Routes = props => (
   <ApolloProvider store={props.store} client={client} >
@@ -64,6 +72,12 @@ const Routes = props => (
           </Route>
           <Route path="budget">
             <IndexRoute component={BudgetSummary} />
+            <Route path="details" component={BudgetSummaryDetails}></Route>
+            <Route path="summaryUse" component={SummaryUse}></Route>
+            <Route path="summaryDepartments" component={SummaryDepartments}></Route>
+            <Route path="summaryCashFlow" component={SummaryCashFlow}></Route>
+            <Route path="detailsTreemap" component={BudgetDetailsTreemap}></Route>
+            <Route path="detailsTable" component={BudgetDetailsTable}></Route>
           </Route>
           <Route path="maintenance">
             <IndexRoute component={Maintenance} />
